+++ react/components/Dropdown/index.js
@@ -97,6 +97,8 @@ ${this.getDropdownIdentification()}`)
     let containerClasses = 'br2 bw1 relative '
     let selectClasses = 'o-0 absolute top-0 left-0 w-100 bottom-0 '
 
+    let labelClasses = 'vtex-dropdown__label dib mb3 w-100 '
+
     const valueLabel = this.getValueLabel()
     const showCaption = !valueLabel
 
@@ -109,16 +111,19 @@ ${this.getDropdownIdentification()}`)
       case 'large':
         classes += 'f5 pv4 pl6 pr5 '
         selectClasses += 'f5 '
+        labelClasses += 'f5 '
         iconSize = 18
         break
       case 'x-large':
         classes += 'f4 pv5 pl7 pr6 '
         selectClasses += 'f4 '
+        labelClasses += 'f5 '
         iconSize = 22
         break
       default:
         classes += 'f6 pv3 pl5 pr4 '
         selectClasses += 'f6 '
+        labelClasses += 'f6 '
         iconSize = 16
         break
     }
@@ -145,7 +150,7 @@ ${this.getDropdownIdentification()}`)
       <div className="vtex-dropdown">
         <label>
           {label && (
-            <span className="vtex-dropdown__label dib mb3 w-100">{label}</span>
+            <span className={labelClasses}>{label}</span>
           )}
           <div className={containerClasses} style={containerStyle}>
             <div id={id} className={`vtex-dropdown__button ${classes}`}>